LEP-4637
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@12461 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/tools/servicebuilder/ServiceBuilder.java
@@ -1966,14 +1966,16 @@ public class ServiceBuilder {
 
 		writeFile(ejbFile, content);
 
-		ejbFile = new File(
-			_outputPath + "/service/persistence/" + entity.getName() +
-				"Persistence.java");
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/persistence/" + entity.getName() +
+					"Persistence.java");
 
-		if (ejbFile.exists()) {
-			System.out.println("Relocating " + ejbFile);
+			if (ejbFile.exists()) {
+				System.out.println("Relocating " + ejbFile);
 
-			ejbFile.delete();
+				ejbFile.delete();
+			}
 		}
 	}
 
@@ -2035,14 +2037,16 @@ public class ServiceBuilder {
 
 		writeFile(ejbFile, content);
 
-		ejbFile = new File(
-			_outputPath + "/service/persistence/" + entity.getName() +
-				"Util.java");
+		if (!_serviceOutputPath.equals(_outputPath)) {
+			ejbFile = new File(
+				_outputPath + "/service/persistence/" + entity.getName() +
+					"Util.java");
 
-		if (ejbFile.exists()) {
-			System.out.println("Relocating " + ejbFile);
+			if (ejbFile.exists()) {
+				System.out.println("Relocating " + ejbFile);
 
-			ejbFile.delete();
+				ejbFile.delete();
+			}
 		}
 	}
